requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==4.4.2
Django==4.1.3
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-multiselectfield==0.1.12
django-timezone-field==5.0
django-widget-tweaks==1.4.12
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
imageio==2.22.4
imageio-ffmpeg==0.4.7
kombu==5.2.4
MouseInfo==0.1.3
moviepy==1.0.0
numpy==1.23.5
outcome==1.2.0
Pillow==9.3.0
proglog==0.1.10
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytweening==1.0.4
pytz==2022.7.1
redis==4.4.2
requests==2.28.1
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.6
wsproto==1.2.0
instagrapi==1.17.1
uwsgi==2.0.21

selenium-stealth
beautifulsoup4
requests
cryptography
lxml
fake-useragent